Move creation of MapPins into ViewModel
