+++ gnd/src/main/java/com/google/android/gnd/ui/home/mapcontainer/MapContainerFragment.java
@@ -18,8 +18,6 @@ package com.google.android.gnd.ui.home.mapcontainer;
 
 import static com.google.android.gnd.rx.RxAutoDispose.autoDisposable;
 import static com.google.android.gnd.rx.RxAutoDispose.disposeOnDestroy;
-import static com.google.android.gnd.util.ImmutableSetCollector.toImmutableSet;
-import static java8.util.stream.StreamSupport.stream;
 
 import android.content.res.ColorStateList;
 import android.os.Bundle;
@@ -38,7 +36,6 @@ import com.google.android.gnd.MainViewModel;
 import com.google.android.gnd.R;
 import com.google.android.gnd.databinding.MapContainerFragBinding;
 import com.google.android.gnd.model.Project;
-import com.google.android.gnd.model.feature.Feature;
 import com.google.android.gnd.rx.BooleanOrError;
 import com.google.android.gnd.rx.Loadable;
 import com.google.android.gnd.system.PermissionsManager.PermissionDeniedException;
@@ -47,10 +44,8 @@ import com.google.android.gnd.ui.common.AbstractFragment;
 import com.google.android.gnd.ui.home.FeatureSheetState;
 import com.google.android.gnd.ui.home.HomeScreenViewModel;
 import com.google.android.gnd.ui.map.MapAdapter;
-import com.google.android.gnd.ui.map.MapPin;
 import com.google.android.gnd.ui.map.MapProvider;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
-import com.google.common.collect.ImmutableSet;
 import io.reactivex.Single;
 import javax.inject.Inject;
 
@@ -130,7 +125,7 @@ public class MapContainerFragment extends AbstractFragment {
   private void onMapReady(MapAdapter map) {
     Log.d(TAG, "MapAdapter ready. Updating subscriptions");
     // Observe events emitted by the ViewModel.
-    mapContainerViewModel.getFeatures().observe(this, features -> setMapPins(map, features));
+    mapContainerViewModel.getMapPins().observe(this, map::setMapPins);
     mapContainerViewModel
         .getLocationLockState()
         .observe(this, state -> onLocationLockStateChange(state, map));
@@ -144,21 +139,6 @@ public class MapContainerFragment extends AbstractFragment {
     addFeatureBtn.setOnClickListener(
         __ -> homeScreenViewModel.onAddFeatureBtnClick(map.getCameraTarget()));
     enableLocationLockBtn();
-  }
-
-  // TODO: Move to ViewModel.
-  private void setMapPins(MapAdapter map, ImmutableSet<Feature> features) {
-    map.setMapPins(stream(features).map(MapContainerFragment::toMapPin).collect(toImmutableSet()));
-  }
-
-  // TODO: Move to ViewModel.
-  private static MapPin toMapPin(Feature feature) {
-    return MapPin.newBuilder()
-        .setId(feature.getId())
-        .setPosition(feature.getPoint())
-        .setStyle(feature.getLayer().getDefaultStyle())
-        .setFeature(feature)
-        .build();
   }
 
   @Override
